Refactorizacion funcionando pero mucho codigo repetido
