+++ src/exceptions/PropiedadNoPuedeConstruirHotelException.java
@@ -1,0 +1,4 @@
+package exceptions;
+
+public class PropiedadNoPuedeConstruirHotelException extends RuntimeException {
+}
