+++ src/modelo/tablero/tipos_casilleros/Barrio.java
@@ -10,11 +10,16 @@ public abstract class Barrio extends Casillero implements Edificable {
 
     private Jugador propietario;
     private int precioTerreno;
-    private int cantidadCasas;
-    private int cantidadHoteles;
-    private int precioCasa;
-    private int precioHotel;
-    private Hashtable<Integer, Integer> alquileres;
+
+
+
+    public Barrio(int nuevoPrecio){
+        if (nuevoPrecio < 0)
+            throw new PrecioNegativoException();
+        this.precioTerreno = nuevoPrecio;
+        this.propietario = null;
+
+    }
 
 
     public Jugador getPropietario() {
@@ -25,28 +30,18 @@ public abstract class Barrio extends Casillero implements Edificable {
 
     public int getPrecio() { return this.precioTerreno; }
 
-    public void pisar(Jugador jugador) {
-        if ((jugador != this.propietario) && (this.propietario != null)) {
-            jugador.pagar(alquileres.get(cantidadCasas));
-        } //aca hay que refactorizar, para que el pisar de barrio verifique, las cantidad de casas, si tiene propietario
-        //asi cuando alguien que que no es propietaario le saquen dinero
-    }
+
 
     @Override
-    public int getCantidadDeCasas() {
-        return this.cantidadCasas;
-    }
+    public abstract int getCantidadDeCasas();
 
     @Override
-    public int getCantidadDeHoteles() {
-        return this.cantidadHoteles;
-    }
-
+    public abstract int getCantidadDeHoteles();
     @Override
     public abstract void agregarCasa();
 
+
     @Override
     public abstract void agregarHotel();
-
 
 }
